requirements.txt
streamlit==1.35.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.0
scikit-learn==1.6.1
pillow==10.3.0

